Revert "openstack-cloudkitty-ui: failed to build a9ed2c5"
This reverts commit 82b6de5a9aa2999930acbf84290c442e4dfefa7a.

Reason for revert: This was centos9 ftbfs review related to:

DEBUG: ERROR: Package 'cloudkitty-dashboard' requires a different Python: 3.9.21 not in '>=3.10'

It passed ci because centos9 jobs are not longer voting in rpm-master branch but the content of the change is only adding a #removeme line in the changelog that should not be merged. Given that in centos10 there is no ftbfs https://trunk.rdoproject.org/centos10-master/status_report.html and in centos9 there is no suitable fix, the best way would be to just abandon it or keep just opened, as new ones will be created on each new commit.

Change-Id: I6b5acbea0b761266d23ecac21e6c4a8921568a5c